refactor: split bundle sync code into source and target
